Tree Working Complete
Tree Working Complete

+++ DSA Project/Classes/Graph.cs
@@ -28,8 +28,8 @@ namespace DSA_Project.Classes
             {
                 int src = edge.Src;
                 int dest = edge.Dest;
-                int lat = edge.src_point;
-                int lng = edge.dest_point;
+                double lat = edge.src_point;
+                double lng = edge.dest_point;
 
                 Node newNode = new Node
                 {
@@ -49,8 +49,8 @@ namespace DSA_Project.Classes
         {
             int src = edge.Src;
             int dest = edge.Dest;
-            int src_point = edge.src_point;
-            int dest_point = edge.dest_point;
+            double src_point = edge.src_point;
+            double dest_point = edge.dest_point;
 
             Node newNode = new Node
             {
@@ -100,8 +100,8 @@ namespace DSA_Project.Classes
     {
         public int Dest { get; set; }
         public Node Next { get; set; }
-        public int src_point { get; set; }
-        public int dest_point { get; set; }
+        public double src_point { get; set; }
+        public double dest_point { get; set; }
     }
 
     // Create Graph Edges / Connection
@@ -110,8 +110,8 @@ namespace DSA_Project.Classes
 
         public int Src { get; set; }
         public int Dest { get; set; }
-        public int src_point { get; set; }
-        public int dest_point { get; set; }
+        public double src_point { get; set; }
+        public double dest_point { get; set; }
     }
